Detect non-call uses of forbidden functions
The code base may not call a forbidden function directly but take a
pointer and pass it around. Make sure that the plugin also catches such
uses. This requires not only iterating over all GIMPLE statements but
also walking all their operands.

Depending on the kind of access, print either 'call to forbidden
function' or 'use of forbidden function'.